backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
PyMuPDF==1.24.5
Pillow==10.3.0
pytesseract==0.3.13
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
distro==1.9.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.8.1
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0

